adding new demos to docs
