+++ docs/src/components/Demo/index.js
@@ -3,6 +3,7 @@
 import React, { Component } from 'react';
 import draftToHtml from 'draftjs-to-html';
 import { Editor } from 'react-draft-wysiwyg';
+import uploadImageCallBack from '../../util/uploadImageCallBack';
 import styles from './styles.css'; // eslint-disable-line no-unused-vars
 
 export default class Demo1 extends Component {
@@ -17,47 +18,99 @@ export default class Demo1 extends Component {
     });
   };
 
-  uploadImageCallBack: Function = (file) => new Promise(
-      (resolve, reject) => {
-        const xhr = new XMLHttpRequest(); // eslint-disable-line no-undef
-        xhr.open('POST', 'https://api.imgur.com/3/image');
-        xhr.setRequestHeader('Authorization', 'Client-ID 8d26ccd12712fca');
-        const data = new FormData(); // eslint-disable-line no-undef
-        data.append('image', file);
-        xhr.send(data);
-        xhr.addEventListener('load', () => {
-          const response = JSON.parse(xhr.responseText);
-          resolve(response);
-        });
-        xhr.addEventListener('error', () => {
-          const error = JSON.parse(xhr.responseText);
-          reject(error);
-        });
-      }
-    );
-
   render() {
     const { editorContent } = this.state;
     return (
       <div className="demo-root">
+        <div className="demo-label">
+          Editor with output generated in HTML.
+        </div>
         <div className="demo-editorSection">
-          <div className="demo-editorWrapper">
-            <Editor
-              toolbarClassName="demo-toolbar"
-              wrapperClassName="demo-wrapper"
-              editorClassName="demo-editor"
-              onChange={this.onEditorChange}
-              toolbarAlwaysVisible
-              uploadImageCallBack={this.uploadImageCallBack}
-            />
-          </div>
+          <Editor
+            toolbarClassName="demo-toolbar"
+            wrapperClassName="demo-wrapper"
+            editorClassName="demo-editor"
+            onChange={this.onEditorChange}
+            toolbar={{
+              image: {
+                uploadCallback: uploadImageCallBack,
+              },
+            }}
+          />
           <textarea
+            disabled
             className="demo-content no-focus"
             value={draftToHtml(editorContent)}
           />
         </div>
         <div className="demo-label">
-          Documentation is work in progress.
+          Editor with output generated in JSON.
+        </div>
+        <div className="demo-editorSection">
+          <Editor
+            toolbarClassName="demo-toolbar"
+            wrapperClassName="demo-wrapper"
+            editorClassName="demo-editor"
+            onChange={this.onEditorChange}
+            toolbar={{
+              image: {
+                uploadCallback: uploadImageCallBack,
+              },
+            }}
+          />
+          <textarea
+            disabled
+            className="demo-content no-focus"
+            value={JSON.stringify(editorContent, null, 4)}
+          />
+        </div>
+        <div className="demo-label">
+          Editor with similar options grouped in drop-down.
+        </div>
+        <div className="demo-editorSection">
+          <Editor
+            toolbarClassName="demo-toolbar"
+            wrapperClassName="demo-wrapper-wide"
+            editorClassName="demo-editor"
+            onChange={this.onEditorChange}
+            toolbar={{
+              inline: {
+                inDropdown: true,
+              },
+              list: {
+                inDropdown: true,
+              },
+              textAlign: {
+                inDropdown: true,
+              },
+              link: {
+                inDropdown: true,
+              },
+              image: {
+                uploadCallback: uploadImageCallBack,
+              },
+              history: {
+                inDropdown: true,
+              },
+            }}
+          />
+        </div>
+        <div className="demo-label">
+          Editor with only a sub-set of options available.
+        </div>
+        <div className="demo-editorSection">
+          <Editor
+            toolbarClassName="demo-toolbar"
+            wrapperClassName="demo-wrapper-wide"
+            editorClassName="demo-editor"
+            onChange={this.onEditorChange}
+            toolbar={{
+              options: ['inline', 'blockType', 'fontSize', 'fontFamily', 'remove', 'history'],
+              inline: {
+                options: ['bold', 'italic', 'underline', 'strikethrough'],
+              },
+            }}
+          />
         </div>
       </div>
     );
