+++ docs/src/components/Demo/styles.css
@@ -10,18 +10,24 @@
   display: block;
   font-size: 20px;
   font-weight: 500;
-  margin: 30px 0;
+  margin: 50px 0;
 }
 .demo-editorSection {
   display: flex;
   justify-content: space-between;
 }
 .demo-wrapper {
-  width: 700px;
+  width: 500px;
   display: block;
+  margin-bottom: 50px;
+}
+.demo-wrapper-wide {
+  width: 715px;
+  display: block;
+  margin-bottom: 50px;
 }
 .demo-editor {
-  height: 400px;
+  height: 300px;
   border: 1px solid #F1F1F1;
   padding: 5px;
   border-radius: 2px;
@@ -35,8 +41,9 @@
   border: 1px solid #F1F1F1;
   padding: 5px;
   border-radius: 5px;
-  width: 400px;
+  width: 300px;
   border-radius: 2px;
-  height: 380px;
-  margin: 100px 0 0 30px;
+  height: 425px;
+  margin-left: 30px;
+  resize: none;
 }
